system_monitor/requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
msgpack==1.1.0
psutil==6.1.0
python-decouple==3.8
redis==5.2.1
sqlparse==0.5.2
tzdata==2024.2
